+++ tg_bot/modules/sql/users_sql.py
@@ -18,7 +18,7 @@
 
 import threading
 
-from sqlalchemy import Column, Integer, UnicodeText, String, ForeignKey, UniqueConstraint, func
+from sqlalchemy import Column, Integer, UnicodeText, String, ForeignKey, UniqueConstraint, func, BigInteger
 
 from tg_bot import dispatcher
 from tg_bot.modules.sql import BASE, SESSION
@@ -26,7 +26,7 @@ from tg_bot.modules.sql import BASE, SESSION
 
 class Users(BASE):
     __tablename__ = "users"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     username = Column(UnicodeText)
 
     def __init__(self, user_id, username=None):
@@ -59,7 +59,7 @@ class ChatMembers(BASE):
                              onupdate="CASCADE",
                              ondelete="CASCADE"),
                   nullable=False)
-    user = Column(Integer,
+    user = Column(BigInteger,
                   ForeignKey("users.user_id",
                              onupdate="CASCADE",
                              ondelete="CASCADE"),
@@ -122,7 +122,7 @@ def update_user(user_id, username, chat_id=None, chat_name=None):
         member = SESSION.query(ChatMembers).filter(ChatMembers.chat == chat.chat_id,
                                                    ChatMembers.user == user.user_id).first()
         if not member:
-            chat_member = ChatMembers(chat.chat_id, user.user_id)
+            chat_member = ChatMembers(int(chat.chat_id), user.user_id)
             SESSION.add(chat_member)
 
         SESSION.commit()
