+++ tg_bot/modules/sql/warns_sql.py
@@ -18,7 +18,7 @@
 
 import threading
 
-from sqlalchemy import Integer, Column, String, UnicodeText, func, distinct, Boolean
+from sqlalchemy import Integer, Column, String, UnicodeText, func, distinct, Boolean, BigInteger
 from sqlalchemy.dialects import postgresql
 
 from tg_bot.modules.sql import SESSION, BASE
@@ -27,7 +27,7 @@ from tg_bot.modules.sql import SESSION, BASE
 class Warns(BASE):
     __tablename__ = "warns"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14), primary_key=True)
     num_warns = Column(Integer, default=0)
     reasons = Column(postgresql.ARRAY(UnicodeText))
